+++ src/Grisette/Internal/SymPrim/Prim/Term.hs
@@ -20,6 +20,7 @@ module Grisette.Internal.SymPrim.Prim.Term
 where
 
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.BVPEval ()
+import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalBitCastTerm ()
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalBitwiseTerm ()
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalDivModIntegralTerm
 import Grisette.Internal.SymPrim.Prim.Internal.Instances.PEvalFP